fix tests references
