+++ remote-messaging/remote-messaging-impl/src/test/java/com/duckduckgo/remote/messaging/newtab/RemoteMessageViewModelTest.kt
@@ -21,12 +21,16 @@ import app.cash.turbine.test
 import com.duckduckgo.common.test.CoroutineTestRule
 import com.duckduckgo.common.utils.playstore.PlayStoreUtils
 import com.duckduckgo.remote.messaging.api.Action.DefaultBrowser
+import com.duckduckgo.remote.messaging.api.Action.Survey
 import com.duckduckgo.remote.messaging.api.Content
 import com.duckduckgo.remote.messaging.api.RemoteMessage
 import com.duckduckgo.remote.messaging.api.RemoteMessageModel
 import com.duckduckgo.remote.messaging.impl.newtab.RemoteMessageViewModel
+import com.duckduckgo.remote.messaging.impl.newtab.RemoteMessageViewModel.Command.SubmitUrl
+import com.duckduckgo.survey.api.SurveyParameterManager
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.test.runTest
+import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Before
@@ -45,6 +49,7 @@ class RemoteMessageViewModelTest {
     private var mockLifecycleOwner: LifecycleOwner = mock()
     private var remoteMessageModel: RemoteMessageModel = mock()
     private var playStoreUtils: PlayStoreUtils = mock()
+    private var surveyParameterManager: SurveyParameterManager = mock()
 
     @Before
     fun setUp() {
@@ -52,6 +57,7 @@ class RemoteMessageViewModelTest {
             coroutineRule.testDispatcherProvider,
             remoteMessageModel,
             playStoreUtils,
+            surveyParameterManager,
         )
     }
 
@@ -139,6 +145,25 @@ class RemoteMessageViewModelTest {
             verify(remoteMessageModel).onActionClicked(remoteMessage)
             val command = awaitItem()
             assertTrue(command is RemoteMessageViewModel.Command.LaunchDefaultBrowser)
+        }
+    }
+
+    @Test
+    fun whenMessageActionClickedIsSurveyThenSubmitUrlCommandSent() = runTest {
+        val remoteMessage = RemoteMessage("id1", Content.Small("", ""), emptyList(), emptyList())
+        whenever(surveyParameterManager.buildSurveyUrl("https://example.com", listOf("atb"))).thenReturn("https://example.com?atb")
+        whenever(remoteMessageModel.getActiveMessages()).thenReturn(flowOf(remoteMessage))
+        whenever(remoteMessageModel.onPrimaryActionClicked(remoteMessage)).thenReturn(
+            Survey("https://example.com", mapOf("queryParams" to "atb")),
+        )
+        testee.onStart(mockLifecycleOwner)
+
+        testee.commands().test {
+            testee.onMessagePrimaryButtonClicked()
+            verify(remoteMessageModel).onPrimaryActionClicked(remoteMessage)
+            val command = awaitItem()
+            assertTrue(command is SubmitUrl)
+            assertEquals("https://example.com?atb", (command as SubmitUrl).url)
         }
     }
